Introduce PersistentCollectionTrait
